requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.6.2
charset-normalizer==3.3.2
dj-database-url==2.2.0
dj-email-url==1.0.6
Django==5.0.6
django-cache-url==3.4.5
django-localflavor==4.0
environs==11.0.0
gunicorn==22.0.0
idna==3.7
marshmallow==3.21.3
packaging==24.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-stdnum==1.20
requests==2.32.3
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
whitenoise==6.7.0
